Rename imaages from Landing and Involved components
